chatbot-server/requirements.txt
fastapi
uvicorn
sqlalchemy
sqlalchemy-utils 
databases
mysql-connector-python
python-dotenv
httpx
pymysql
google-generativeai
genai
pydantic-settings
cryptography
pinecone-client
python-docx 
PyPDF2 
pandas 
openpyxl
python-multipart
# openai
bcrypt 
pyjwt
fastapi-mail
cloudinary
beautifulsoup4 
lxml


